+++ libcoveofds/additionalfields.py
@@ -8,7 +8,7 @@ class AdditionalFields:
 
     def process(self, json_data: dict) -> list:
 
-        schema_fields = self._schema.get_fields()
+        schema_fields = self._schema.get_package_schema_fields()
 
         additional_fields_all = get_additional_fields_info(json_data, schema_fields)